refactor, move utilities into common folder
